refactor(aether-actor): invert Local/cost storage via an injected backend
Make `aether-actor` target-blind in its storage logic by inverting the
`Local` backend: the crate defines the trait + one shared `ActorSlots` and a
process-global provider install slot, and whoever drives the actor installs
the "current slots" backend. The host (`aether-substrate`) installs a
`thread_local!`-routed backend at boot; a single-actor image (the wasm guest)
installs the static backend from its `export!` init prologue. The two
cfg-split `local.rs` backends and trait defs collapse into one (unified on
`Send`), and `CostCell`/`CostCells` — pure host machinery — move to
`aether-substrate` next to `CostTable`.

No `target_arch`/cfg remains in the storage or dispatch logic; the FFI ABI
shims (the genuine wasm boundary) keep their target gate.

Closes #2070.
